Add Mocha tests for utils
