+++ scripts/utils.js
@@ -1,0 +1,232 @@
+// Utility functions for text analysis
+
+const stopWords = [
+  'a',
+  'à',
+  'ao',
+  'aos',
+  'aquela',
+  'aquelas',
+  'aquele',
+  'aqueles',
+  'aquilo',
+  'as',
+  'às',
+  'até',
+  'com',
+  'como',
+  'da',
+  'das',
+  'de',
+  'dela',
+  'delas',
+  'dele',
+  'deles',
+  'depois',
+  'do',
+  'dos',
+  'e',
+  'é',
+  'ela',
+  'elas',
+  'ele',
+  'eles',
+  'em',
+  'entre',
+  'era',
+  'eram',
+  'éramos',
+  'essa',
+  'essas',
+  'esse',
+  'esses',
+  'esta',
+  'está',
+  'estamos',
+  'estão',
+  'estar',
+  'estas',
+  'estava',
+  'estavam',
+  'estávamos',
+  'este',
+  'esteja',
+  'estejam',
+  'estejamos',
+  'estes',
+  'esteve',
+  'estive',
+  'estivemos',
+  'estiver',
+  'estivera',
+  'estiveram',
+  'estivéramos',
+  'estiverem',
+  'estivermos',
+  'estivesse',
+  'estivessem',
+  'estivéssemos',
+  'estou',
+  'eu',
+  'foi',
+  'fomos',
+  'for',
+  'fora',
+  'foram',
+  'fôramos',
+  'forem',
+  'formos',
+  'fosse',
+  'fossem',
+  'fôssemos',
+  'fui',
+  'há',
+  'haja',
+  'hajam',
+  'hajamos',
+  'hão',
+  'havemos',
+  'haver',
+  'hei',
+  'houve',
+  'houvemos',
+  'houver',
+  'houvera',
+  'houverá',
+  'houveram',
+  'houvéramos',
+  'houverão',
+  'houverei',
+  'houverem',
+  'houveremos',
+  'houveria',
+  'houveriam',
+  'houveríamos',
+  'houvermos',
+  'houvesse',
+  'houvessem',
+  'houvéssemos',
+  'isso',
+  'isto',
+  'já',
+  'lhe',
+  'lhes',
+  'mais',
+  'mas',
+  'me',
+  'mesmo',
+  'meu',
+  'meus',
+  'minha',
+  'minhas',
+  'muito',
+  'na',
+  'não',
+  'nas',
+  'nem',
+  'no',
+  'nos',
+  'nós',
+  'nossa',
+  'nossas',
+  'nosso',
+  'nossos',
+  'num',
+  'numa',
+  'o',
+  'os',
+  'ou',
+  'para',
+  'pela',
+  'pelas',
+  'pelo',
+  'pelos',
+  'por',
+  'qual',
+  'quando',
+  'que',
+  'quem',
+  'são',
+  'se',
+  'seja',
+  'sejam',
+  'sejamos',
+  'sem',
+  'ser',
+  'será',
+  'serão',
+  'serei',
+  'seremos',
+  'seria',
+  'seriam',
+  'seríamos',
+  'seu',
+  'seus',
+  'só',
+  'somos',
+  'sou',
+  'sua',
+  'suas',
+  'também',
+  'te',
+  'tem',
+  'tém',
+  'temos',
+  'tenha',
+  'tenham',
+  'tenhamos',
+  'tenho',
+  'terá',
+  'terão',
+  'terei',
+  'teremos',
+  'teria',
+  'teriam',
+  'teríamos',
+  'teu',
+  'teus',
+  'teve',
+  'tinha',
+  'tinham',
+  'tínhamos',
+  'tive',
+  'tivemos',
+  'tiver',
+  'tivera',
+  'tiveram',
+  'tivéramos',
+  'tiverem',
+  'tivermos',
+  'tivesse',
+  'tivessem',
+  'tivéssemos',
+  'tu',
+  'tua',
+  'tuas',
+  'um',
+  'uma',
+  'você',
+  'vocês',
+  'vos'
+]
+
+const punct = [',', '"', "'", '.', '!']
+
+const stopWords_ = [...stopWords, ...punct]
+
+function collectTokenSet (hinario) {
+  const tokens = hinario.hinos.reduce((a, h) => {
+    if (h.tokens && h.tokens.pt) return [...a, ...h.tokens.pt]
+    return a
+  }, [])
+  const lower = tokens.map(t => t.toLowerCase())
+  return new Set(lower.filter(t => !stopWords_.includes(t)))
+}
+
+function jaccard (setA, setB) {
+  const inter = [...setA].filter(x => setB.has(x))
+  const union = new Set([...setA, ...setB])
+  return inter.length / union.size
+}
+
+module.exports = { collectTokenSet, jaccard, stopWords_, stopWords, punct }
